fix: use string templates for nest marker
